Addressed change requests - New package and classes for the request types - Shortened the short option for no vacation entries to one letter - Renames inappropriate occurrences of 'Holiday' to 'Vacation'
