+++ src/main/java/main/UserInput.java
@@ -2,7 +2,10 @@
 package main;
 
 import i18n.ResourceHandler;
-import lombok.Getter;
+import main.request.GenerateRequest;
+import main.request.HelpRequest;
+import main.request.Request;
+import main.request.VersionRequest;
 import org.apache.commons.cli.*;
 import org.apache.commons.io.FilenameUtils;
 
@@ -33,7 +36,9 @@ public class UserInput {
 	}
 
 	/**
-	 * Parse the command line arguments
+	 * Parse the command line arguments and returns the request read from the
+	 * command line arguments. This request will also contain additional parameters
+	 * or information, like excluding vacation entries on a generate request.
 	 * 
 	 * @return The user request read from the command line arguments
 	 * @throws ParseException Thrown in case the command line arguments are invalid
@@ -53,7 +58,7 @@ public class UserInput {
 		if (commandLine.hasOption(UserInputOption.GUI.getOption().getOpt()) && commandLine.hasOption(UserInputOption.FILE.getOption().getOpt())) {
 			throw new ParseException(ResourceHandler.getMessage("error.userinput.mutuallyExclusiveOptionsGuiFile"));
 		} else {
-			return new GenerateRequest(commandLine.hasOption(UserInputOption.NO_HOLIDAY.getOption().getOpt()));
+			return new GenerateRequest(commandLine.hasOption(UserInputOption.NO_VACATION_ENTRIES.getOption().getOpt()));
 		}
 	}
 
@@ -215,66 +220,4 @@ public class UserInput {
 
 		return file;
 	}
-
-	/**
-	 * The general type of action a user requested through the command line
-	 * arguments
-	 */
-	public enum RequestType {
-		HELP, VERSION, GENERATE
-	}
-
-	/**
-	 * Base class for an action a user requested through the command line arguments.
-	 * Contains only a request type, and cannot be instantiated outside of this
-	 * class.
-	 */
-	@Getter
-	public static class Request {
-		private final RequestType type;
-
-		private Request(RequestType type) {
-			this.type = type;
-		}
-	}
-
-	/**
-	 * A help request. The Request to be returned if the user has requested
-	 * help.<br/>
-	 * The {@code getType()} method will return {@link RequestType#HELP}.
-	 */
-	public static class HelpRequest extends Request {
-		public HelpRequest() {
-			super(RequestType.HELP);
-		}
-	}
-
-	/**
-	 * A version request. The Request to be returned if the user has requested the
-	 * version.<br/>
-	 * The {@code getType()} method will return {@link RequestType#VERSION}.
-	 */
-	@Getter
-	public static class VersionRequest extends Request {
-		public VersionRequest() {
-			super(RequestType.VERSION);
-		}
-	}
-
-	/**
-	 * A generation request. The Request to be returned if the user has requested to
-	 * generate a time sheet. In this case, the user can specify if they want to
-	 * explicitly generate holiday entries or not.<br/>
-	 * The {@code getType()} method will return {@link RequestType#GENERATE}.
-	 */
-	@Getter
-	public static class GenerateRequest extends Request {
-		private final boolean excludeHolidayEntries;
-
-		public GenerateRequest(boolean excludeHolidayEntries) {
-			super(RequestType.GENERATE);
-			this.excludeHolidayEntries = excludeHolidayEntries;
-		}
-	}
-
 }
