+++ src/main/java/main/request/RequestType.java
@@ -1,0 +1,10 @@
+/* Licensed under MIT 2025. */
+package main.request;
+
+/**
+ * The general type of action a user requested through the command line
+ * arguments
+ */
+public enum RequestType {
+	HELP, VERSION, GENERATE
+}
