+++ src/main/resources/i18n/MessageBundle.properties
@@ -3,7 +3,7 @@ application.name = TimeSheetGenerator
 command.input.help.description = Prints helping information
 command.input.version.description = Prints the version
 command.input.gui.description = Enables load/save dialogs
-command.input.no-holiday.description = If enabled, will not add explicit entries for holiday entries
+command.input.no-vacation.description = If enabled, will not add explicit entries for vacation entries
 command.input.file.description = Passes file paths via console
 command.input.file.arguments = global.json> <month.json> <output-file.tex
 command.output.version = Version: {0}
